pyproject.toml
[project]
name = "local-read-write-agent"
version = "0.1.0"
description = "A Langchain agent that manages fridge inventory and suggests recipes based on user preferences"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-ollama>=0.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
